changed the simulation logic

+++ Testing/ESP32 Tests/Protocolo/include/utils.h
@@ -1,4 +1,4 @@
 #include <Arduino.h>
 #include <stdint.h>
 
-String array2String(double* arr, uint16_t arr_len, String separator=",");
+String array2String(int* arr, uint16_t arr_len, String separator=",");
