+++ Testing/ESP32 Tests/Protocolo/src/utils.cpp
@@ -1,6 +1,6 @@
 #include "utils.h"
 
-String array2String(double* arr, uint16_t arr_len, String separator){
+String array2String(int* arr, uint16_t arr_len, String separator){
     uint8_t i;
     String s = String(arr[0]);
     for (i = 1; i < arr_len; i++){
